refactor and correct codes
