+++ pages/[id].js
@@ -1,5 +1,8 @@
 import ServiceCard from "@/components/ServiceCard";
+import { useRouter } from "next/router";
 
 export default function ServiceDetailsPage() {
-  return <ServiceCard />;
+  const router = useRouter();
+  const { id } = router.query;
+  return <ServiceCard id={id} />;
 }
